requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.21.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.2.0
anthropic>=0.30.0
APScheduler>=3.10.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
